#56 feat: GroupRepository 추가 및 sql문 추가
