+++ motimo-domain/src/main/java/kr/co/domain/group/repository/GroupMemberRepository.java
@@ -1,0 +1,4 @@
+package kr.co.domain.group.repository;
+
+public interface GroupMemberRepository {
+}
