+++ motimo-domain/src/main/java/kr/co/domain/group/repository/GroupRepository.java
@@ -1,0 +1,5 @@
+package kr.co.domain.group.repository;
+
+public interface GroupRepository {
+
+}
