+++ motimo-infra-rdb/src/main/java/kr/co/infra/rdb/group/entity/GroupEntity.java
@@ -1,12 +1,14 @@
-package kr.co.infra.rdb.group;
+package kr.co.infra.rdb.group.entity;
 
+import jakarta.persistence.CascadeType;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EntityListeners;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.Id;
-import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import java.time.LocalDateTime;
+import java.util.List;
 import java.util.UUID;
 import kr.co.infra.rdb.common.entity.BaseEntity;
 import kr.co.infra.rdb.common.uuid.GeneratedUuidV7Value;
@@ -18,23 +20,17 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 @Entity
 @EntityListeners(AuditingEntityListener.class)
-@Table(name = "group_users")
+@Table(name = "groups")
 @SoftDelete(columnName = "is_deleted")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class GroupMemberEntity extends BaseEntity {
+public class GroupEntity extends BaseEntity {
     @Id
     @GeneratedUuidV7Value
     private UUID id;
 
-    private UUID userId;
+    @OneToMany(fetch = FetchType.LAZY, mappedBy = "group", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<GroupMemberEntity> groupMembers;
 
-    private UUID goalId;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    private GroupEntity group;
-
-    private LocalDateTime joinedDate;
-
-    private boolean isNotificationActive;
+    private LocalDateTime finishedDate;
 }
