+++ motimo-infra-rdb/src/main/java/kr/co/infra/rdb/group/entity/GroupMemberEntity.java
@@ -1,14 +1,12 @@
-package kr.co.infra.rdb.group;
+package kr.co.infra.rdb.group.entity;
 
-import jakarta.persistence.CascadeType;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EntityListeners;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
+import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import java.time.LocalDateTime;
-import java.util.List;
 import java.util.UUID;
 import kr.co.infra.rdb.common.entity.BaseEntity;
 import kr.co.infra.rdb.common.uuid.GeneratedUuidV7Value;
@@ -20,17 +18,23 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 @Entity
 @EntityListeners(AuditingEntityListener.class)
-@Table(name = "groups")
+@Table(name = "group_users")
 @SoftDelete(columnName = "is_deleted")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class GroupEntity extends BaseEntity {
+public class GroupMemberEntity extends BaseEntity {
     @Id
     @GeneratedUuidV7Value
     private UUID id;
 
-    @OneToMany(fetch = FetchType.LAZY, mappedBy = "group", cascade = CascadeType.ALL, orphanRemoval = true)
-    private List<GroupMemberEntity> groupMembers;
+    private UUID userId;
 
-    private LocalDateTime finishedDate;
+    private UUID goalId;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private GroupEntity group;
+
+    private LocalDateTime joinedDate;
+
+    private boolean isNotificationActive;
 }
