+++ motimo-infra-rdb/src/main/java/kr/co/infra/rdb/group/reposiotry/GroupJpaRepository.java
@@ -1,0 +1,8 @@
+package kr.co.infra.rdb.group.reposiotry;
+
+import java.util.UUID;
+import kr.co.infra.rdb.group.entity.GroupEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface GroupJpaRepository extends JpaRepository<GroupEntity, UUID> {
+}
