+++ motimo-infra-rdb/src/main/java/kr/co/infra/rdb/group/reposiotry/GroupMemberRepositoryImpl.java
@@ -1,0 +1,11 @@
+package kr.co.infra.rdb.group.reposiotry;
+
+import kr.co.domain.group.repository.GroupMemberRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+
+@Repository
+@RequiredArgsConstructor
+public class GroupMemberRepositoryImpl implements GroupMemberRepository {
+    private final GroupMemberJpaRepository groupMemberJpaRepository;
+}
