+++ motimo-infra-rdb/src/main/java/kr/co/infra/rdb/group/reposiotry/GroupRepositoryImpl.java
@@ -1,0 +1,14 @@
+package kr.co.infra.rdb.group.reposiotry;
+
+import kr.co.domain.group.repository.GroupRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+
+@Repository
+@RequiredArgsConstructor
+public class GroupRepositoryImpl implements GroupRepository {
+
+    private final GroupJpaRepository groupJpaRepository;
+
+
+}
